测试引入断言失败判断
